requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
psycopg2-binary 
passlib[bcrypt]
python-jose
python-decouple